打折

+++ mabhmad/model/catalog/product.php
@@ -708,53 +708,73 @@ $sql = "INSERT INTO " . DB_PREFIX . "product SET product_id = '" . $this->db->es
         }
 
 			
-		
-		if (isset($data['product_special'])) {
-			//是否同步同类产品的折扣信息
-		
-			foreach ($data['product_special'] as $key => $product_special) {
-			
-                    if (isset($product_special['product_special_id'])){
+
+// print_r($product_id);
+
+// print_r($data['product_special']);
+$sp_id=array();
+        
+                if (isset($data['product_special'])) {
+            //是否同步同类产品的折扣信息
+        
+            foreach ($data['product_special'] as $key => $product_special) {
+            
+                    if (isset($product_special['product_special_id'])&&$product_special['product_special_id']>0){
+                        $sp_id[]=$product_special['product_special_id'];
                         if (isset($product_special['date_end']) && !empty($product_special['date_end'])) {
                            $this->db->query("UPDATE " . DB_PREFIX . "product_special SET
-					        customer_group_id = '" . (int)$product_special['customer_group_id'] . "',
-					        	product_option_value_id = '" . (int)$product_special['product_option_value_id'] . "',
-					        priority = '" . (int)$product_special['priority'] . "',
-					        price = '" . (float)$product_special['price'] . "',
-					        percent = '" . (float)$product_special['percent'] . "',
-					        date_start = '" . $this->db->escape($product_special['date_start']) . "',
-					        date_end = '" . $this->db->escape($product_special['date_end']) . "' WHERE product_special_id = '" . $product_special['product_special_id'] . "'");
+                            customer_group_id = '" . (int)$product_special['customer_group_id'] . "',
+                                product_option_value_id = '" . (int)$product_special['product_option_value_id'] . "',
+                            priority = '" . (int)$product_special['priority'] . "',
+                            price = '" . (float)$product_special['price'] . "',
+                            percent = '" . (float)$product_special['percent'] . "',
+                            date_start = '" . $this->db->escape($product_special['date_start']) . "',
+                            date_end = '" . $this->db->escape($product_special['date_end']) . "' WHERE product_special_id = '" . $product_special['product_special_id'] . "'");
                         }
                         else {
                          $this->db->query("DELETE FROM " . DB_PREFIX . "product_special WHERE product_special_id = '" . $product_special['product_special_id'] . "'");
                         }
                     }
                     else {
-                    	 if (isset($product_special['date_end']) && !empty($product_special['date_end'])) {
+                         if (isset($product_special['date_end']) && !empty($product_special['date_end'])) {
+                          
                        $this->db->query("INSERT INTO " . DB_PREFIX . "product_special SET
-					    product_id = '" . (int)$product_id . "',
-					    customer_group_id = '" . (int)$product_special['customer_group_id'] . "',
-					    priority = '" . (int)$product_special['priority'] . "',
-					    	product_option_value_id = '" . (int)$product_special['product_option_value_id'] . "',
-					    price = '" . (float)$product_special['price'] . "',
-					    percent = '" . (float)$product_special['percent'] . "',
-					    date_start = '" . $this->db->escape($product_special['date_start']) . "',
-					    date_end = '" . $this->db->escape($product_special['date_end']) . "'");
-         //               print_r("INSERT INTO " . DB_PREFIX . "product_special SET
-					    // product_id = '" . (int)$product_id . "',
-					    // customer_group_id = '" . (int)$product_special['customer_group_id'] . "',
-					    // priority = '" . (int)$product_special['priority'] . "',
-					    // 	product_option_value_id = '" . (int)$product_special['product_option_value_id'] . "',
-					    // price = '" . (float)$product_special['price'] . "',
-					    // percent = '" . (float)$product_special['percent'] . "',
-					    // date_start = '" . $this->db->escape($product_special['date_start']) . "',
-					    // date_end = '" . $this->db->escape($product_special['date_end']) . "'");exit();
-                   		}
+                        product_id = '" . (int)$product_id . "',
+                        customer_group_id = '" . (int)$product_special['customer_group_id'] . "',
+                        priority = '" . (int)$product_special['priority'] . "',
+                            product_option_value_id = '" . (int)$product_special['product_option_value_id'] . "',
+                        price = '" . (float)$product_special['price'] . "',
+                        percent = '" . (float)$product_special['percent'] . "',
+                        date_start = '" . $this->db->escape($product_special['date_start']) . "',
+                        date_end = '" . $this->db->escape($product_special['date_end']) . "'");
+                       $sp_id[]= $this->db->getLastId();
+                        }
                     }
-				
-		
-		}
-	}
+                
+        
+        }
+        if ( $sp_id) {
+           $this->db->query("DELETE FROM " . DB_PREFIX . "product_special WHERE product_id = '" . (int)$product_id . "' AND product_special_id not in(".implode(',',  $sp_id).")");
+        }else{
+              $this->db->query("DELETE FROM " . DB_PREFIX . "product_special WHERE product_id = '" . (int)$product_id . "'");
+                       
+        }
+    }
+
+    // 同步副站
+
+ $q=  $this->db->query("SELECT * FROM " . DB_PREFIX . "product_special WHERE product_id = '" . (int)$product_id . "' ");
+
+ 	$this->querysql("DELETE FROM " . DB_PREFIX . "product_special WHERE product_id = '" . (int)$product_id . "'");
+	 if ( $q->rows) {
+	 	foreach ($q->rows as $key => $row) {
+	 		$this->copsql('product_special',$row);
+	 	}
+	 	
+	 }
+
+
+    // 
 
 		$price=$this->getProductSpecials($product_id);
 		// print_r($price);exit();
